requirements.txt
cffi==1.17.1
cryptography==45.0.6
noiseprotocol==0.3.1
pycparser==2.22
